requirements.txt
cloudpickle==2.0.0
cycler==0.11.0
fonttools==4.28.2
gym==0.21.0
gym-retro==0.8.0
imageio==2.13.0
importlib-metadata==4.8.2
keyboard==0.13.5
kiwisolver==1.3.2
matplotlib==3.5.0
Multi-Template-Matching==1.6.3
networkx==2.6.3
numpy==1.21.4
opencv-python==4.5.4.60
opencv-python-headless==4.5.4.60
packaging==21.3
pandas==1.3.4
Pillow==8.4.0
pyglet==1.5.21
pyparsing==3.0.6
python-dateutil==2.8.2
pytz==2021.3
PyWavelets==1.2.0
scikit-image==0.18.3
scipy==1.7.3
setuptools-scm==6.3.2
six==1.16.0
tifffile==2021.11.2
tomli==1.2.2
typing-extensions==4.0.0
zipp==3.6.0
